+++ datalab/gui/actionhandler.py
@@ -638,6 +638,27 @@ class BaseActionHandler(metaclass=abc.ABCMeta):
 
         with self.new_category(ActionCategory.EDIT):
             self.new_action(
+                _("Recompute"),
+                icon_name="recompute.svg",
+                shortcut="Ctrl+R",
+                tip=_("Recompute selected %s with its processing parameters")
+                % self.OBJECT_STR,
+                triggered=self.panel.recompute_processing,
+                select_condition=SelectCond.at_least_one_group_or_one_object,
+                context_menu_pos=-1,
+                toolbar_pos=-1,
+            )
+            self.new_action(
+                _("Select source objects"),
+                icon_name="goto_source.svg",
+                tip=_("Select source object(s) used to create the selected %s")
+                % self.OBJECT_STR,
+                triggered=self.panel.select_source_objects,
+                select_condition=SelectCond.exactly_one,
+                context_menu_pos=-1,
+                toolbar_pos=-1,
+            )
+            self.new_action(
                 _("Rename"),
                 icon_name="rename.svg",
                 shortcut="F2",
@@ -645,6 +666,7 @@ class BaseActionHandler(metaclass=abc.ABCMeta):
                 triggered=self.panel.rename_selected_object_or_group,
                 select_condition=SelectCond.exactly_one_group_or_one_object,
                 context_menu_pos=-1,
+                separator=True,
             )
             self.new_action(
                 _("New group..."),
@@ -653,7 +675,6 @@ class BaseActionHandler(metaclass=abc.ABCMeta):
                 triggered=self.panel.new_group,
                 select_condition=SelectCond.always,
                 context_menu_pos=-1,
-                toolbar_pos=-1,
             )
             self.new_action(
                 _("Move up"),